Prevent primary "data" resources from showing up again in the "included" array
This would occur in a situation where the primary resource has an included relationship which in turn has a relationship back to the primary resource. Ideally we would merge the "included" resource into the primary "data" resource, but I don't think it will really be a problem in the real-world, and would probably require some significant reworking of the Document class, so I haven't bothered for now.

The original bug was discovered by axdr here: https://discuss.flarum.org/d/2156-prepareapiattributes-is-being-fired-for-one-discussion-50-times