Better support for empty relationships. closes #94
To return an empty to-one relationship, pass an EmptyResource instance as the Relationship data.
To return an empty to-many relationship, pass an EmptyCollection instance as the Relationship data. (Passing a Collection with an empty array will work too.)

Why necessitate an EmptyResource object rather than just passing null as the Relationship data? Because there are times where the user may not wish to include the `data` key in the relationship object at all, instead filling it with `links` or `meta` – null data value is reserved for this case.